Hotfix/fridge post

+++ users/views.py
@@ -395,7 +395,12 @@ class UserDetailView(APIView):
 class UserDetailFridgeView(APIView):
     permission_classes = [IsAuthenticated]
 
-    def get(self, request):
+    def get(self, request, **kwargs):
+        if kwargs.get("fridge_id"):
+            fridge = get_object_or_404(Fridge, id=kwargs.get("fridge_id"))
+            serializer = UserFridgeSerializer(fridge)
+            return Response(serializer.data, status=status.HTTP_200_OK)
+
         all_fridge = Fridge.objects.filter(user=request.user)
         serializer = UserFridgeSerializer(
             all_fridge,
@@ -403,19 +408,32 @@ class UserDetailFridgeView(APIView):
         )
         return Response(serializer.data, status=status.HTTP_200_OK)
 
-    def post(self, request):
+    def post(self, request, **kwargs):
+        if kwargs.get("fridge_id"):
+            return Response(
+                {"error": "올바르지않은 요청입니다"}, status=status.HTTP_405_METHOD_NOT_ALLOWED
+            )
         serializer = UserFridgeSerializer(data=request.data)
         if serializer.is_valid():
             fridge = serializer.save(
                 user=request.user,
+                ingredient_id=request.get("ingredient"),
             )
             serializer = UserFridgeSerializer(fridge)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         else:
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    def delete(self, request, fridge_id):
+    def delete(self, request, **kwargs):
+        try:
+            fridge_id = kwargs["fridge_id"]
+        except:
+            return Response(
+                {"error": "올바르지않은 요청입니다."}, status=status.HTTP_405_METHOD_NOT_ALLOWED
+            )
         fridges = get_object_or_404(Fridge, pk=fridge_id)
+        if fridges.user != request.user:
+            raise PermissionDenied
         if fridges:
             fridges.delete()
             return Response({"message": "삭제완료"}, status=status.HTTP_200_OK)
